Extract state builder test fixtures
Move the fixture-only helper constructors out of
state_builders_composition_unit.rs into a local support module.

This keeps the main integration test under the 400-line limit
without leaking those helpers into the shared backend/tests/support
facade, which would trigger dead-code warnings in unrelated test
crates.
